refactor: organize generation
