+++ src/services/workflowGenerationService.ts
@@ -10,18 +10,26 @@ import {
   GenerationPromptResponse,
   GenerationSession,
 } from "@/types/generation-session";
-import type {
-  NodeType,
-  WorkflowConnection,
-  WorkflowDefinition,
-  WorkflowNode,
-  WorkflowExecution,
-} from "@/types/workflow";
-import { generateId } from "@/utils/helpers";
+import type { WorkflowDefinition } from "@/types/workflow";
+import { generateId, newWorkflowId } from "@/utils/helpers";
 import {
   buildWorkflowNodeTools,
   getNodeSchema,
 } from "./workflowGenerationNodeTools";
+import {
+  connectNodes,
+  createNode,
+  deleteNode,
+  disconnectNodes,
+  getLatestExecution,
+  inferNodeType,
+  moveNode,
+  setUrlPattern,
+  setWorkflowDescription,
+  setWorkflowEnabled,
+  setWorkflowName,
+  updateNodeConfig,
+} from "./workflowTools";
 
 const openrouter = createOpenRouter({
   apiKey:
@@ -35,19 +43,9 @@ const model = openrouter.chat("openai/gpt-oss-120b:free", {
   maxTokens: 2000,
 });
 
-const triggerNodeTypes = new Set([
-  "page-load",
-  "component-load",
-  "delay",
-  "key-press",
-  "http-request",
-  "button-click",
-  "popup",
-]);
-
 function createBaseWorkflow(): WorkflowDefinition {
   return {
-    id: generateId("workflow", 12),
+    id: newWorkflowId(),
     name: "Untitled workflow",
     description: "",
     urlPattern: "https://*",
@@ -59,81 +57,13 @@ function createBaseWorkflow(): WorkflowDefinition {
   };
 }
 
-function ensureDraftWorkflow(session: GenerationSession): WorkflowDefinition {
-  return session.draftWorkflow ?? createBaseWorkflow();
-}
-
-function isRecord(value: unknown): value is Record<string, any> {
-  return Boolean(value) && typeof value === "object" && !Array.isArray(value);
-}
-
-function asString(value: unknown): string | undefined {
-  return typeof value === "string" && value.trim().length > 0
-    ? value.trim()
-    : undefined;
-}
-
-function getFirstString(
-  args: Record<string, any>,
-  keys: string[],
-): string | undefined {
-  for (const key of keys) {
-    const value = asString(args[key]);
-    if (value) {
-      return value;
-    }
+function ensureWorkflow(session: GenerationSession): WorkflowDefinition {
+  if (!session.workflowId) {
+    return createBaseWorkflow();
   }
-
-  return undefined;
-}
-
-function asPosition(value: unknown): { x: number; y: number } | undefined {
-  if (
-    isRecord(value) &&
-    typeof value.x === "number" &&
-    typeof value.y === "number"
-  ) {
-    return { x: value.x, y: value.y };
-  }
-
-  return undefined;
-}
-
-function nodeTypeFromArgs(args: Record<string, any>): NodeType | null {
-  const explicit = getFirstString(args, [
-    "nodeType",
-    "node_type",
-    "category",
-    "kind",
-  ]);
-  return explicit === "trigger" || explicit === "action" ? explicit : null;
-}
-
-function inferNodeType(type: string, args: Record<string, any>): NodeType {
-  const explicit = nodeTypeFromArgs(args);
-  if (explicit) {
-    return explicit;
-  }
-
-  return triggerNodeTypes.has(type) ? "trigger" : "action";
-}
-
-function getNextNodePosition(nodes: WorkflowNode[]): { x: number; y: number } {
-  if (nodes.length === 0) {
-    return { x: 280, y: 120 };
-  }
-
-  const rightmostX = nodes.reduce(
-    (max, node) => Math.max(max, node.position.x),
-    nodes[0].position.x,
+  return (
+    useStore.getState().readWorkflow(session.workflowId) || createBaseWorkflow()
   );
-  const averageY =
-    nodes.reduce((sum, node) => sum + node.position.y, 0) / nodes.length;
-
-  return {
-    x: rightmostX + 240,
-    y: Math.max(80, Math.round(averageY)),
-  };
 }
 
 function appendMessage(
@@ -188,10 +118,15 @@ function historyToPrompt(messages: GenerationMessage[]): string {
 function summarizeValue(value: any): string {
   if (value === null) return "null";
   if (value === undefined) return "undefined";
-  if (typeof value === "string") return JSON.stringify(value.length > 80 ? `${value.slice(0, 77)}...` : value);
-  if (typeof value === "number" || typeof value === "boolean") return String(value);
+  if (typeof value === "string")
+    return JSON.stringify(
+      value.length > 80 ? `${value.slice(0, 77)}...` : value,
+    );
+  if (typeof value === "number" || typeof value === "boolean")
+    return String(value);
   if (Array.isArray(value)) return `[Array(${value.length})]`;
-  if (typeof value === "object") return `{${Object.keys(value).slice(0, 8).join(", ")}${Object.keys(value).length > 8 ? ", ..." : ""}}`;
+  if (typeof value === "object")
+    return `{${Object.keys(value).slice(0, 8).join(", ")}${Object.keys(value).length > 8 ? ", ..." : ""}}`;
   return String(value);
 }
 
@@ -220,7 +155,7 @@ function buildNodeVariableReference(): string {
 }
 
 function buildSystemPrompt(session: GenerationSession): string {
-  const workflow = ensureDraftWorkflow(session);
+  const workflow = ensureWorkflow(session);
   const pageContext = session.pageContext
     ? JSON.stringify(session.pageContext, null, 2)
     : "null";
@@ -256,7 +191,7 @@ function buildSystemPrompt(session: GenerationSession): string {
     "Use getNodeSchema when you need the exact config fields for a node type.",
     `Available node types: ${JSON.stringify(availableNodeTypes)}`,
     "",
-    `Current draft workflow: ${JSON.stringify(workflow, null, 2)}`,
+    `Current workflow: ${JSON.stringify(workflow, null, 2)}`,
     `Current page context: ${pageContext}`,
     `Recent conversation: ${history || "(empty)"}`,
     "",
@@ -272,69 +207,14 @@ function commitSession(
 ): GenerationSession {
   const updatedSession = {
     ...nextSession,
-    draftWorkflow: nextWorkflow,
+    workflowId: nextWorkflow.id,
     updatedAt: Date.now(),
-  };
+  } as GenerationSession;
 
   useStore.getState().updateWorkflow(nextWorkflow);
   useStore.getState().setActiveGenerationSession(updatedSession);
 
   return updatedSession;
-}
-
-function getLatestExecution(workflowId: string): WorkflowExecution | null {
-  const executions = useStore.getState().executions;
-  for (let index = executions.length - 1; index >= 0; index -= 1) {
-    if (executions[index]?.workflowId === workflowId) {
-      return executions[index];
-    }
-  }
-
-  return null;
-}
-
-function setWorkflowName(
-  workflow: WorkflowDefinition,
-  name: string,
-): WorkflowDefinition {
-  return { ...workflow, name, modifiedAt: Date.now() };
-}
-
-function setWorkflowDescription(
-  workflow: WorkflowDefinition,
-  description: string,
-): WorkflowDefinition {
-  return { ...workflow, description, modifiedAt: Date.now() };
-}
-
-function setUrlPattern(
-  workflow: WorkflowDefinition,
-  urlPattern: string,
-): WorkflowDefinition {
-  return { ...workflow, urlPattern, modifiedAt: Date.now() };
-}
-
-function setWorkflowEnabled(
-  workflow: WorkflowDefinition,
-  enabled: boolean,
-): WorkflowDefinition {
-  return { ...workflow, enabled, modifiedAt: Date.now() };
-}
-
-function deriveWorkflowName(prompt: string): string {
-  const words = prompt
-    .trim()
-    .replace(/\s+/g, " ")
-    .split(" ")
-    .filter(Boolean)
-    .slice(0, 6)
-    .join(" ");
-
-  if (!words) {
-    return "Generated workflow";
-  }
-
-  return words.length > 48 ? `${words.slice(0, 45).trimEnd()}...` : words;
 }
 
 function finalizeGeneratedWorkflow(
@@ -354,294 +234,20 @@ function finalizeGeneratedWorkflow(
   return nextWorkflow;
 }
 
-function createNode(
-  workflow: WorkflowDefinition,
-  args: Record<string, any>,
-): { workflow: WorkflowDefinition; result: string } {
-  const type = getFirstString(args, ["type", "actionType", "triggerType"]);
-  if (!type) {
-    throw new Error("createNode requires type");
+function deriveWorkflowName(prompt: string): string {
+  const words = prompt
+    .trim()
+    .replace(/\s+/g, " ")
+    .split(" ")
+    .filter(Boolean)
+    .slice(0, 6)
+    .join(" ");
+
+  if (!words) {
+    return "Generated workflow";
   }
 
-  const nodeType = inferNodeType(type, args);
-  const nodeId =
-    getFirstString(args, ["id", "nodeId", "node_id"]) ||
-    generateId(nodeType, 8);
-  const position =
-    asPosition(args.position) || getNextNodePosition(workflow.nodes);
-  const config = isRecord(args.config) ? args.config : {};
-  const inputs =
-    Array.isArray(args.inputs) && args.inputs.length > 0
-      ? args.inputs.filter(
-          (value): value is string => typeof value === "string",
-        )
-      : nodeType === "trigger"
-        ? []
-        : ["input"];
-  const outputs =
-    Array.isArray(args.outputs) && args.outputs.length > 0
-      ? args.outputs.filter(
-          (value): value is string => typeof value === "string",
-        )
-      : ["output"];
-
-  const node: WorkflowNode = {
-    id: nodeId,
-    type: nodeType,
-    position,
-    data: { type, config },
-    inputs,
-    outputs,
-  };
-
-  return {
-    workflow: {
-      ...workflow,
-      nodes: [...workflow.nodes, node],
-      modifiedAt: Date.now(),
-    },
-    result: `Created ${nodeType} node '${type}' (${nodeId}).`,
-  };
-}
-
-function updateNodeConfig(
-  workflow: WorkflowDefinition,
-  args: Record<string, any>,
-): { workflow: WorkflowDefinition; result: string } {
-  const nodeId = getFirstString(args, [
-    "nodeId",
-    "node_id",
-    "targetNodeId",
-    "target_node_id",
-    "id",
-  ]);
-  if (!nodeId) {
-    throw new Error("updateNodeConfig requires nodeId");
-  }
-
-  const patch = isRecord(args.config)
-    ? args.config
-    : isRecord(args.patch)
-      ? args.patch
-      : null;
-  if (!patch) {
-    throw new Error("updateNodeConfig requires config or patch");
-  }
-
-  let updated = false;
-  const nodes = workflow.nodes.map((node) => {
-    if (node.id !== nodeId) {
-      return node;
-    }
-
-    updated = true;
-    return {
-      ...node,
-      data: {
-        ...node.data,
-        config: {
-          ...(node.data as any).config,
-          ...patch,
-        },
-      },
-    };
-  });
-
-  if (!updated) {
-    throw new Error(`Node '${nodeId}' not found`);
-  }
-
-  return {
-    workflow: {
-      ...workflow,
-      nodes,
-      modifiedAt: Date.now(),
-    },
-    result: `Updated config for node '${nodeId}'.`,
-  };
-}
-
-function moveNode(
-  workflow: WorkflowDefinition,
-  args: Record<string, any>,
-): { workflow: WorkflowDefinition; result: string } {
-  const nodeId = getFirstString(args, [
-    "nodeId",
-    "node_id",
-    "targetNodeId",
-    "target_node_id",
-    "id",
-  ]);
-  const position = asPosition(args.position);
-
-  if (!nodeId) {
-    throw new Error("moveNode requires nodeId");
-  }
-  if (!position) {
-    throw new Error("moveNode requires position { x, y }");
-  }
-
-  let updated = false;
-  const nodes = workflow.nodes.map((node) => {
-    if (node.id !== nodeId) {
-      return node;
-    }
-
-    updated = true;
-    return { ...node, position };
-  });
-
-  if (!updated) {
-    throw new Error(`Node '${nodeId}' not found`);
-  }
-
-  return {
-    workflow: {
-      ...workflow,
-      nodes,
-      modifiedAt: Date.now(),
-    },
-    result: `Moved node '${nodeId}'.`,
-  };
-}
-
-function deleteNode(
-  workflow: WorkflowDefinition,
-  args: Record<string, any>,
-): { workflow: WorkflowDefinition; result: string } {
-  const nodeId = getFirstString(args, [
-    "nodeId",
-    "node_id",
-    "targetNodeId",
-    "target_node_id",
-    "id",
-  ]);
-  if (!nodeId) {
-    throw new Error("deleteNode requires nodeId");
-  }
-
-  if (!workflow.nodes.some((node) => node.id === nodeId)) {
-    throw new Error(`Node '${nodeId}' not found`);
-  }
-
-  return {
-    workflow: {
-      ...workflow,
-      nodes: workflow.nodes.filter((node) => node.id !== nodeId),
-      connections: workflow.connections.filter(
-        (connection) =>
-          connection.source !== nodeId && connection.target !== nodeId,
-      ),
-      modifiedAt: Date.now(),
-    },
-    result: `Deleted node '${nodeId}'.`,
-  };
-}
-
-function connectNodes(
-  workflow: WorkflowDefinition,
-  args: Record<string, any>,
-): { workflow: WorkflowDefinition; result: string } {
-  const sourceId = getFirstString(args, [
-    "sourceId",
-    "source_id",
-    "from",
-    "source",
-  ]);
-  const targetId = getFirstString(args, [
-    "targetId",
-    "target_id",
-    "to",
-    "target",
-  ]);
-  const sourceHandle = getFirstString(args, ["sourceHandle", "source_handle"]);
-  const targetHandle = getFirstString(args, ["targetHandle", "target_handle"]);
-
-  if (!sourceId || !targetId) {
-    throw new Error("connectNodes requires sourceId and targetId");
-  }
-
-  const connection: WorkflowConnection = {
-    id:
-      getFirstString(args, ["id", "connectionId", "connection_id"]) ||
-      generateId("conn", 8),
-    source: sourceId,
-    target: targetId,
-    sourceHandle,
-    targetHandle,
-  };
-
-  const exists = workflow.connections.some(
-    (item) =>
-      item.source === connection.source &&
-      item.target === connection.target &&
-      item.sourceHandle === connection.sourceHandle &&
-      item.targetHandle === connection.targetHandle,
-  );
-
-  if (exists) {
-    return {
-      workflow,
-      result: `Connection from '${sourceId}' to '${targetId}' already exists.`,
-    };
-  }
-
-  return {
-    workflow: {
-      ...workflow,
-      connections: [...workflow.connections, connection],
-      modifiedAt: Date.now(),
-    },
-    result: `Connected '${sourceId}' to '${targetId}'.`,
-  };
-}
-
-function disconnectNodes(
-  workflow: WorkflowDefinition,
-  args: Record<string, any>,
-): { workflow: WorkflowDefinition; result: string } {
-  const connectionId = getFirstString(args, [
-    "connectionId",
-    "connection_id",
-    "id",
-  ]);
-  const sourceId = getFirstString(args, [
-    "sourceId",
-    "source_id",
-    "from",
-    "source",
-  ]);
-  const targetId = getFirstString(args, [
-    "targetId",
-    "target_id",
-    "to",
-    "target",
-  ]);
-
-  if (!connectionId && (!sourceId || !targetId)) {
-    throw new Error(
-      "disconnectNodes requires connectionId or sourceId and targetId",
-    );
-  }
-
-  return {
-    workflow: {
-      ...workflow,
-      connections: workflow.connections.filter((connection) => {
-        if (connectionId) {
-          return connection.id !== connectionId;
-        }
-
-        return !(
-          connection.source === sourceId && connection.target === targetId
-        );
-      }),
-      modifiedAt: Date.now(),
-    },
-    result: connectionId
-      ? `Disconnected connection '${connectionId}'.`
-      : `Disconnected '${sourceId}' from '${targetId}'.`,
-  };
+  return words.length > 48 ? `${words.slice(0, 45).trimEnd()}...` : words;
 }
 
 export class WorkflowGenerationService {
@@ -661,7 +267,7 @@ export class WorkflowGenerationService {
     const { session, prompt } = request;
     const cleanedSession = stripPendingThinkingMessage(session);
 
-    let workingWorkflow = ensureDraftWorkflow(cleanedSession);
+    let workflow = ensureWorkflow(cleanedSession);
     let workingSession = cleanedSession;
 
     const persist = (
@@ -674,14 +280,14 @@ export class WorkflowGenerationService {
         message,
         kind,
       );
-      commitSession(workingWorkflow, workingSession);
+      commitSession(workflow, workingSession);
     };
 
     const applyWorkflow = (
       nextWorkflow: WorkflowDefinition,
       message: string,
     ) => {
-      workingWorkflow = nextWorkflow;
+      workflow = nextWorkflow;
       persist(message, "tool");
     };
 
@@ -696,7 +302,7 @@ export class WorkflowGenerationService {
             description: "Set the workflow name.",
             inputSchema: z.object({ name: z.string().min(1) }),
             execute: async ({ name }) => {
-              const nextWorkflow = setWorkflowName(workingWorkflow, name);
+              const nextWorkflow = setWorkflowName(workflow, name);
               applyWorkflow(nextWorkflow, `Set workflow name to '${name}'.`);
               return { name };
             },
@@ -706,7 +312,7 @@ export class WorkflowGenerationService {
             inputSchema: z.object({ description: z.string() }),
             execute: async ({ description }) => {
               const nextWorkflow = setWorkflowDescription(
-                workingWorkflow,
+                workflow,
                 description,
               );
               applyWorkflow(nextWorkflow, "Updated workflow description.");
@@ -717,7 +323,7 @@ export class WorkflowGenerationService {
             description: "Set the workflow URL pattern.",
             inputSchema: z.object({ urlPattern: z.string().min(1) }),
             execute: async ({ urlPattern }) => {
-              const nextWorkflow = setUrlPattern(workingWorkflow, urlPattern);
+              const nextWorkflow = setUrlPattern(workflow, urlPattern);
               applyWorkflow(
                 nextWorkflow,
                 `Set URL pattern to '${urlPattern}'.`,
@@ -729,7 +335,7 @@ export class WorkflowGenerationService {
             description: "Enable or disable the workflow.",
             inputSchema: z.object({ enabled: z.boolean() }),
             execute: async ({ enabled }) => {
-              const nextWorkflow = setWorkflowEnabled(workingWorkflow, enabled);
+              const nextWorkflow = setWorkflowEnabled(workflow, enabled);
               applyWorkflow(
                 nextWorkflow,
                 `Set workflow enabled to ${enabled}.`,
@@ -744,7 +350,7 @@ export class WorkflowGenerationService {
               type: z.string().min(1),
             }),
             execute: async ({ type }) => {
-              const kind = triggerNodeTypes.has(type) ? "trigger" : "action";
+              const kind = inferNodeType(type, {});
               const schema = getNodeSchema(kind, type);
 
               if (!schema) {
@@ -771,7 +377,7 @@ export class WorkflowGenerationService {
               "Get the latest stored execution for the workflow currently being edited.",
             inputSchema: z.object({}),
             execute: async () => {
-              const workflowId = workingSession.draftWorkflow?.id;
+              const workflowId = workingSession.workflowId;
               if (!workflowId) {
                 return { execution: null };
               }
@@ -780,16 +386,15 @@ export class WorkflowGenerationService {
             },
           }),
           ...buildWorkflowNodeTools({
-            getWorkflow: () => workingWorkflow,
-            createNode: (args) => createNode(workingWorkflow, args),
+            getWorkflow: () => workflow,
+            createNode: (args) => createNode(workflow, args),
             commitWorkflow: (nextWorkflow, message) => {
-              workingWorkflow = nextWorkflow;
+              workflow = nextWorkflow;
               persist(message, "tool");
               workingSession = {
                 ...workingSession,
-                draftWorkflow: nextWorkflow,
               };
-              commitSession(workingWorkflow, workingSession);
+              commitSession(workflow, workingSession);
             },
           }),
           updateNodeConfig: tool({
@@ -804,7 +409,7 @@ export class WorkflowGenerationService {
             }),
             execute: async (args) => {
               const { workflow: nextWorkflow, result: message } =
-                updateNodeConfig(workingWorkflow, args);
+                updateNodeConfig(workflow, args);
               applyWorkflow(nextWorkflow, message);
               return { message };
             },
@@ -820,7 +425,7 @@ export class WorkflowGenerationService {
             }),
             execute: async (args) => {
               const { workflow: nextWorkflow, result: message } = moveNode(
-                workingWorkflow,
+                workflow,
                 args,
               );
               applyWorkflow(nextWorkflow, message);
@@ -837,7 +442,7 @@ export class WorkflowGenerationService {
             }),
             execute: async (args) => {
               const { workflow: nextWorkflow, result: message } = deleteNode(
-                workingWorkflow,
+                workflow,
                 args,
               );
               applyWorkflow(nextWorkflow, message);
@@ -859,7 +464,7 @@ export class WorkflowGenerationService {
             }),
             execute: async (args) => {
               const { workflow: nextWorkflow, result: message } = connectNodes(
-                workingWorkflow,
+                workflow,
                 args,
               );
               applyWorkflow(nextWorkflow, message);
@@ -881,7 +486,7 @@ export class WorkflowGenerationService {
             }),
             execute: async (args) => {
               const { workflow: nextWorkflow, result: message } =
-                disconnectNodes(workingWorkflow, args);
+                disconnectNodes(workflow, args);
               applyWorkflow(nextWorkflow, message);
               return { message };
             },
@@ -897,13 +502,12 @@ export class WorkflowGenerationService {
         "result",
       );
 
-      workingWorkflow = finalizeGeneratedWorkflow(workingWorkflow, prompt);
+      workflow = finalizeGeneratedWorkflow(workflow, prompt);
       workingSession = {
         ...workingSession,
-        draftWorkflow: workingWorkflow,
       };
 
-      workingSession = commitSession(workingWorkflow, workingSession);
+      workingSession = commitSession(workflow, workingSession);
 
       return { session: workingSession };
     } catch (error) {
@@ -917,7 +521,7 @@ export class WorkflowGenerationService {
         "error",
       );
       workingSession = { ...workingSession, status: "failed" };
-      workingSession = commitSession(workingWorkflow, workingSession);
+      workingSession = commitSession(workflow, workingSession);
 
       return { session: workingSession };
     }
